feat: Add basic content ready for launch
