+++ src/content/pages/index.mdx
@@ -7,6 +7,7 @@ import { css } from '@emotion/core';
 
 import Argument from 'components/Argument';
 import AnimatedGIF from 'components/AnimatedGIF';
+import CTAButtons from 'components/CTAButtons';
 import Panel from 'components/Panel';
 import SignUpForm from 'components/SignUpForm';
 import Spacer from 'components/Spacer';
@@ -40,14 +41,14 @@ Whether you're taking notes in another app, reviewing a prototype in Figma, or j
 
 </TextBlock>
 
-![A Mac showing a microphone icon in the menubar.](images/menubar-muting.png)
+![Mic Drop welcome window.](images/onboarding-welcome.png)
 
 </Panel>
 
 
 <Panel gridTablet="1fr 1fr" gridWide="44rem auto 44rem">
 
-![A Mac showing a microphone icon in the menubar.](images/menubar-muting.png)
+![The Mic Drop preferences window.](images/customisable.png)
 
 <TextBlock end>
 
@@ -74,26 +75,17 @@ Is this thing on? Can you hear me now? A clear indicator means you’ll never as
 
 <TextBlock>
 
-## Mic Drop is currently in beta. Want to try it?
-
-We're looking for beta testers to make sure everything works well before launch. If you'd like to get an invitation to join our beta programme, share your information below and we'll get back to you shortly.
-
-<SignUpForm />
-
-</TextBlock>
-
-<Spacer />
-
-<TextBlock end>
-
 ## Now, get ready for your meetings to suck less.
 
 We can't promise to keep your overly-enthusiastic co-worker from cutting you off, but we _can_ make it a little bit easier to control your sound.
-
-All beta testers will receive a lifetime copy of the app, as a token of our appreciation.
+onboarding-screenshot
+<CTAButtons />
 
 </TextBlock>
 
+![Mic Drop window showing current mute status.](images/onboarding-screenshot.png)
+
+<!--
 <Argument title="No Captain Crunching.">
 Forgot your breakfast? No worries! Crunch your Happy Os to your heart’s content, without driving your co-workers bananas.
 </Argument>
@@ -109,6 +101,7 @@ Forgot your breakfast? No worries! Crunch your Happy Os to your heart’s conten
 <Argument title="No spying." top end>
 There are probably already more than enough machines in your house spying on you. We don’t ever listen in. (Check our privacy policy!)
 </Argument>
+-->
 
 </Panel>
